Fix for retrieving list of interfaces.
Sending in None as reference explicitly tells the DLL we're only interested in the number of interfaces available. I was having crashes with this request otherwise.

Note: I tested and verified this with Vimba 2.0. Perhaps the issue doesn't arise with older versions, but the documentation seems to indicate that this is the intended method all along.