for people who do not want to know what is pymba
This is for people who do not want to know what is pymba, and do not want to know the image format, just want to get a image from AVT camera and start to work with OpenCV, AI.  You can simply import AVTCamera, get the AVT camera by index, read the image. It is base on pymba and opencv, make sure you use PIP to install both module.